move some methods to a new file mode.go
